+++ signal_utils/radar_pulse.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from pathlib import Path
 
 import numpy as np
@@ -9,7 +8,7 @@ from .common.filter_windows import create_fir_filter, nuttall_window
 from .common.generate_bpsk import generate_bpsk
 
 
-def read_input_params(filename: Path) -> tuple[int, Decimal, int, list[int], int, Decimal, int]:
+def read_input_params(filename: Path) -> tuple[int, float, int, list[int], int, float, int]:
     """Read radar pulse parameters from input YAML file
 
     Args:
@@ -24,7 +23,7 @@ def read_input_params(filename: Path) -> tuple[int, Decimal, int, list[int], int
     return input_params['sample_rate'], input_params['bit_length'], input_params['num_bits'], input_params['taps'], \
         input_params['amplitude'], input_params['pri'], input_params['num_pulses']
 
-def generate_pulse(seq: NDArray[np.int_], sample_rate: int, bit_length: Decimal, pri: Decimal, num_pulses: int) -> NDArray[np.complex_]:
+def generate_pulse(seq: NDArray[np.int_], sample_rate: int, bit_length: float, pri: float, num_pulses: int) -> NDArray[np.complex_]:
     samples_per_pulse = int(sample_rate * pri)
 
     pulse = generate_bpsk(seq, sample_rate, bit_length)
